changed delay of int to float

+++ app/ai_workflow.py
@@ -108,7 +108,7 @@ def plot_figure(should_save, should_display, results):
 
     line6, = ax.plot(times, [int(result['ap']) for result in results], label="Person's Average Precision (%)",
                      color=color1, marker='o', markersize=3, linewidth=0)
-    line7, = ax.plot(times, [int((result['delay'])[:-2]) for result in results], label="UE network latency (ms)",
+    line7, = ax.plot(times, [float((result['delay'])[:-2]) for result in results], label="UE network latency (ms)",
                      color=color2, marker='o', markersize=3, linewidth=0)
     # line8, = ax.plot(times, [int((result['rate'][:-4])) for result in results], label="UE network bandwidth (Mbps)",
     #                  color=color3, marker='o', markersize=4)
